+++ lib/fcs_directus.dart
@@ -43,6 +43,8 @@ export 'src/services/users_service.dart';
 export 'src/services/files_service.dart';
 export 'src/services/roles_service.dart';
 export 'src/services/policies_service.dart';
+export 'src/services/activity_service.dart';
+export 'src/services/assets_service.dart';
 
 // Exceptions
 export 'src/exceptions/directus_exception.dart';
@@ -57,6 +59,9 @@ export 'src/models/directus_functions.dart';
 export 'src/models/directus_user.dart';
 export 'src/models/directus_role.dart';
 export 'src/models/directus_policy.dart';
+export 'src/models/directus_activity.dart';
+export 'src/models/directus_revision.dart';
+export 'src/models/asset_transforms.dart';
 
 // WebSocket
 export 'src/websocket/directus_websocket_client.dart';